+++ dbms/src/Interpreters/createBlockSelector.h
@@ -22,6 +22,6 @@ template <typename T>
 IColumn::Selector createBlockSelector(
     const IColumn & column,
     size_t num_shards,
-    const std::vector<size_t> & slots);
+    const std::vector<UInt64> & slots);
 
 }
